Move legacy multicall components to callers module
